Revert Particle.h rename
