+++ firmware/boron-ins-fsm/test/mocks/mock_os_queue_t.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "mock_Particle.h"
+#include "Particle.h"
 #include <stdint.h>
 
 typedef uint32_t system_tick_t;
