+++ firmware/boron-ins-fsm/test/mocks/mock_thread.h
@@ -11,6 +11,5 @@ class Thread
 public:
     Thread() {}
 
-public:
     Thread(const char* name, void (*)(void*)) {}
 };
